+++ app/tag/controller.js
@@ -1,8 +1,21 @@
-const Tag = require('./model')
+const Tag = require('./model');
+
+const { policyFor } = require('../policy/index');
 
 
 async function store(req, res, next) {
     try {
+
+        // ----cek policy -----//
+        let policy = policyFor(req.user);
+
+        if (!policy.can('create', 'Tag')) {
+            return res.json({
+                error: 1,
+                message: 'Anda tidak memiliki akses'
+            })
+        }
+
         let payload = req.body;
 
         const tag = new Tag(payload);
@@ -24,6 +37,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
     try {
+        // ----cek policy -----//
+        let policy = policyFor(req.user);
+
+        if (!policy.can('update', 'Tag')) {
+            return res.json({
+                error: 1,
+                message: 'Anda tidak memiliki akses'
+            })
+        }
         let payload = req.body;
 
         const tag = await Tag.findOneAndUpdate({ _id: req.params.id }, payload, { new: true, runValidators: true })
@@ -43,6 +65,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
     try {
+        // ----cek policy -----//
+        let policy = policyFor(req.user);
+
+        if (!policy.can('delete', 'Tag')) {
+            return res.json({
+                error: 1,
+                message: 'Anda tidak memiliki akses'
+            })
+        }
 
         let deleted = await Tag.findOneAndDelete({ _id: req.params.id });
